Kareem

+++ Scarlet/fav/views.py
@@ -7,7 +7,9 @@ from django.middleware.csrf import get_token
 from django.contrib.auth import authenticate, login
 
 
-class FavCreateView(FormView):
+class FavAlterView(FormView):
+
+    """ alters Favorite instance 'fav/unfav an object' """
 
     form_class = FavoriteForm
     model = Favorite
@@ -17,7 +19,7 @@ class FavCreateView(FormView):
         user = authenticate(username="ahmed", password="1234")
         if user is not None:
             login(self.request, user)
-        fav = form.save(commit=False)
+        fav_value = self.request.POST['fav_value']
         csrf_token_value = get_token(self.request)
         try:
             content_type = ContentType.objects.get(
@@ -25,37 +27,19 @@ class FavCreateView(FormView):
                 model=self.request.POST['model'].lower())
             model_object = content_type.get_object_for_this_type(
                 id=self.request.POST['model_id'])
-            fav.content_object = model_object
-            if fav.user:
-                fav.save()
-            Favorite.objects.get(id=fav.id)
+            if fav_value == 'favorite':
+                fav = form.save(commit=False)
+                fav.content_object = model_object
+                if fav.user:
+                    fav.save()
+                Favorite.objects.get(id=fav.id)
+            else:
+                Favorite.objects.get(
+                    object_id=model_object.id,
+                    user=self.request.user,
+                    content_type=content_type).delete()
         except:
             return JsonResponse({
                 'success': 0,
-                'error': "You have to sign in first"})
-        return JsonResponse({"csrf": csrf_token_value})
-
-
-class FavDeleteView(FormView):
-
-    form_class = FavoriteForm
-    model = Favorite
-    template_name = 'fav/fav_form.html'
-
-    def form_valid(self, form):
-        csrf_token_value = get_token(self.request)
-        try:
-            content_type = ContentType.objects.get(
-                app_label=self.request.POST['app_name'],
-                model=self.request.POST['model'].lower())
-            model_object = content_type.get_object_for_this_type(
-                id=self.request.POST['model_id'])
-            Favorite.objects.get(
-                object_id=model_object.id,
-                user=self.request.user,
-                content_type=content_type).delete()
-        except:
-            return JsonResponse({
-                'success': 0,
-                'error': "You have to sign in first"})
+                'error': "You have to sign in "})
         return JsonResponse({"csrf": csrf_token_value})
